feat: move github calls to service
